compliance-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "compliance-engine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "boto3>=1.34",
    "pyyaml>=6.0",
    "psycopg2-binary>=2.9",
]

[project.scripts]
compliance-engine = "engine.main:main"
